+++ django_mapengine/static/django_mapengine/js/legend.js
@@ -24,10 +24,10 @@ const createLegend = (title, unit, colors, valueRanges, nextColumnStartIndex = 3
     </div>
     <div class="legend__wrap">
       <div class="legend__column">
-        ${valueRanges.filter((value, idx) => idx <= nextColumnStartIndex).map((value, idx) => `<div class="legend__item" id="legend__item__color-${idx}">${value}</div>`).join(' ')}
+        ${valueRanges.filter((value, idx) => idx < nextColumnStartIndex).map((value, idx) => `<div class="legend__item" id="legend__item__color-${idx}">${value}</div>`).join(' ')}
       </div>
       <div class="legend__column">
-        ${valueRanges.filter((value, idx) => idx > nextColumnStartIndex).map((value, idx) => `<div class="legend__item" id="legend__item__color-${idx + nextColumnStartIndex + 1}">${value}</div>`).join(' ')}
+        ${valueRanges.filter((value, idx) => idx >= nextColumnStartIndex).map((value, idx) => `<div class="legend__item" id="legend__item__color-${idx + nextColumnStartIndex}">${value}</div>`).join(' ')}
       </div>
     </div>
     <style>
@@ -75,6 +75,7 @@ function loadLegend(msg, choroplethName){
   }
 
   const entriesPerColumn = Math.floor(valueRanges.length / 2);
+  console.log(entriesPerColumn)
   legendElement.innerHTML = createLegend(title, unit, colors, valueRanges, entriesPerColumn);
   return logMessage(msg);
 }